robot_rpi/requirements.txt
Flask==2.3.0
Flask-SocketIO==5.3.0
python-socketio==5.9.0
eventlet==0.33.3
RPi.GPIO==0.7.1
